Task 5 - Implemented Promotion System
Add a method to convert an employee to a manager when they get a team and attach the employee information with them in their manager role. This is done using the indexOf function to locate the index of the employee info in the array and then move that information into the Manager class.